+++ java-io-base/src/main/java/nbbrd/io/text/TextParser.java
@@ -15,6 +15,7 @@ import java.nio.file.Path;
 import java.util.Optional;
 import java.util.function.Consumer;
 import java.util.function.Function;
+import java.util.stream.Stream;
 
 public interface TextParser<T> {
 
@@ -82,5 +83,10 @@ public interface TextParser<T> {
     @StaticFactoryMethod
     static <T> @NonNull TextParser<T> onParsingReader(@NonNull IOFunction<? super Reader, ? extends T> function) {
         return new FunctionalTextParser<>(function);
+    }
+
+    @StaticFactoryMethod
+    static <T> @NonNull TextParser<T> onParsingLines(@NonNull Function<? super Stream<String>, ? extends T> function) {
+        return new FunctionalTextParser<>(IOFunction.checked(function).compose(FunctionalTextParser::asLines));
     }
 }